@@ -2,7 +2,6 @@ package com.xkcoding.cache.ehcache;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
 
 /**
  * <p>
@@ -13,10 +12,9 @@ import org.springframework.cache.annotation.EnableCaching;
  * @date Created in 2018-11-16 17:02
  */
 @SpringBootApplication
-@EnableCaching
-public class SpringBootDemoCacheEhcacheApplication {
+public class EhcacheCacheApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(SpringBootDemoCacheEhcacheApplication.class, args);
+        SpringApplication.run(EhcacheCacheApplication.class, args);
     }
 }
